==> scripts/04_models_fire_AIM-data.Rmd <==
---
title: "Models of fire occurence at AIM sites"
author: "Martin Holdrege"
date: "`r lubridate::today()`"
output:
  html_document:
    toc: true
    toc_float:
      collapsed: false
---

Modeling fire occurrence (from MTBS, IFPH, and IFPH + MTBS combined), at AIM
sites, with predictor variables being vegetation cover metrics measured
at AIM sites, and Daymet climate data for thos sites. This analysis only includes
AIM sites with >0 sagebrush cover. 

## Dependencies

```{r warning =TRUE, message=FALSE}
library(tidyverse, quietly = TRUE)
library(GGally) # for ggpairs()
library(knitr)
theme_set(theme_classic())

```

# Read in data

This data was downloaded in the 03_download_GEE_output_from_gdrive.R
script, and created in the 02_compile_AIM_data.js script. 

The year range in the file name refers to the time period from which 
fire and climate data is from. 

```{r}
aim1 <- read_csv("data_processed/AIM/AIM-sagebrush-sites_with-climate-and-fire_1985-2019_v1.csv",
                 show_col_types = FALSE)
```

# Exploratory analysis

## Prep data

Columns after 'WoodyHgt_Avg', are ones I added in GEE, the others are part of the
origional AIM dataset available on GEE.

```{r}
names(aim1)
```


```{r}
aim2 <- aim1 %>% 
  mutate( # all count datasets (mtbs and ifph) are 35 years (change if updated)
         mtbs_n = 35, 
         # proportion of years with fires
         mtbs_prop = mtbs/mtbs_n,
         ifph_prop = ifph/mtbs_n,
         comb_prop = comb/mtbs_n)
```

## summary tables

```{r}

aim2 %>% 
  select(comb:comb_prop, -`.geo`) %>% 
  pivot_longer(cols = everything(),
               names_to = 'variable') %>% 
  group_by(variable) %>% 
  summarise(across(value, .fns = list(mean = mean, min = min, 
                                      median = median, max = max))) %>% 
  mutate(across(-variable, .fns = round, digits = 4)) %>% 
  kable(caption = "summary stats of fire and climate variables")
```


## Histograms

Histograms of veg cover variables (potential predictor vars)

```{r, fig.height=8, fig.width=8, message = FALSE}
cover_vars <- names(aim1) %>% 
  str_subset("Cover_AH") %>% 
  str_subset("Tree", negate = TRUE)

grass_vars <- str_subset(cover_vars, "Grass")
other_vars <- str_subset(cover_vars, "Grass", negate = TRUE)
forb_vars <- str_subset(other_vars, "Forb")
other_vars <- str_subset(other_vars, "Forb", negate = TRUE)

print(grass_vars)
ggpairs(aim1[grass_vars],
                diag = list(continuous = 'barDiag')) +
  ggtitle("Grass cover variables")

print(forb_vars)
ggpairs(aim1[forb_vars],
                diag = list(continuous = 'barDiag')) +
  ggtitle("Forb cover variables")

print(other_vars)
ggpairs(aim1[other_vars],
                diag = list(continuous = 'barDiag')) +
  ggtitle("Other cover variables")
  
```

Histograms of potential response variables (fire occurrence)


```{r message = FALSE}
g <- ggplot(aim2) +
  # note the mtbs fir prop goes up to >20%, but they're 
  # are very few cells with that many fires, so truncating
  coord_cartesian(c(0, 10)) +
  labs(x = "Fire probability (%/year)") 

h1 <- g +
  geom_histogram(aes(mtbs_prop*100)) +
  labs(subtitle = "MTBS, fire occurence over 35 years")

h2 <- g +
  geom_histogram(aes(ifph_prop*100)) +
  labs(subtitle = "IFPH, fire occurence over 35 years")

h3 <- g +
  geom_histogram(aes(comb_prop*100)) +
  labs(subtitle = "IFPH and MTBS combined, fire occurence over 35 years")

gridExtra::grid.arrange(h1, h2, h3, ncol = 2)
```

## boxplots

Number of fires occurring vs. plant cover

```{r}

aim2 %>% 
  select(mtbs, all_of(grass_vars)) %>% 
  pivot_longer(cols = all_of(grass_vars)) %>% 
  ggplot(aes(x = as.factor(mtbs), y = value)) +
  geom_boxplot() +
  labs(x = "Number of fires (mtbs)",
       y= "% Cover",
       title = "Grass variables vs fire occurrence") +
  facet_wrap(~name, scales = 'free_y')

aim2 %>% 
  select(mtbs, all_of(forb_vars)) %>% 
  pivot_longer(cols = all_of(forb_vars)) %>% 
  ggplot(aes(x = as.factor(mtbs), y = value)) +
  geom_boxplot() +
  labs(x = "Number of fires (mtbs)",
       y= "% Cover",
       title = "Forb variables vs fire occurrence") +
  facet_wrap(~name, scales = 'free_y')

aim2 %>% 
  select(mtbs, all_of(other_vars)) %>% 
  pivot_longer(cols = all_of(other_vars)) %>% 
  ggplot(aes(x = as.factor(mtbs), y = value)) +
  geom_boxplot() +
  labs(x = "Number of fires (mtbs)",
       y= "% Cover",
       title = "Other variables vs fire occurrence") +
  facet_wrap(~name, scales = 'free_y')

```
