==> scripts/05_models_biome-mask_fire-prob_byNFire.Rmd <==
---
title: "Models of fire probability for the sagebrush biome"
subtitle: "Before/after fire, resampled/stratified data"
author: "Martin Holdrege"
date: "`r lubridate::today()`"
output:
  html_document:
    toc: true
    toc_float:
      collapsed: false
---

This data uses average rap biomass from before a given fire (or for the entire time
period if no fire). So there can be multiple observations (rows) per grid cell (if
there are multiple fires)

Here the training data is resampled to be more balanced (i.e., upsampling
bins of rare data, downsampling bins of common data). 

Note a concern with the up-sampling is that many bins start with only one
datapoint in them, which is then repeated many times in the resampled dataset.
Apparently, having duplicated points like this can increase the risk of 
over-fitting. 

This code is set up to fit models separately to fire data compiled
in a couple of ways (e.g. paint method, vs reduceToImage), and also
multiple fire dataset (e.g. mtbs, ifph etc). Meaning this code
iterates over lists. Right now these lists are of length 1 because I'm only
looking at the cwf (usgs combined wildfire dataset--which is the best there is),
and the paint method to convert polygons to rasters (in GEE)

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning=FALSE, cache.lazy = FALSE,
                      message = FALSE)
```

# Dependencies 

User defined parameters

```{r}
# set to true if want to run for a limited number of rows (i.e. for code testing)
test_run <- FALSE
s <- "_A-P_A2-T2_A2-S2_A-Pr" # string to paste to file names e.g., that defines the interactions in the model
# such as (summer ppt * MAT) and annuals*temperature interactions
```


Sourcing the script that creates a dataframe of historical
fore occurrence, as well as biotic and climate
predictors for those same grid-cells. 

```{r source_data, warning=FALSE, message=FALSE, cache = FALSE}
source("scripts/04_create_biome-mask_dataframe_byNfire.R")
```

```{r warning=FALSE, message=FALSE}
# set option so resampled dataset created here reproduces earlier runs of this code with dplyr 1.0.10
options(dplyr.legacy_locale = TRUE) 
source("src/general_functions.R")
source("src/fig_params.R")
source("src/modeling_functions.R")
source("src/resample_functions.R")
library(tidyverse)
library(GGally) # for ggpairs()
library(pdp) # for partial dependence plots
library(gridExtra)
library(knitr)
library(patchwork) # for figure insets etc. 
theme_set(theme_classic())
```


# Prep data

```{r}
dfs_biome3 <- dfs_byNFire3

# small dataset for if testing the data
if(test_run) {
  dfs_biome3 <- map(dfs_biome3, sample_n, size = 1000)
}
```


```{r prep_data, cache = FALSE}
# only interested in cws data source at the moment (the others are worse
# and make the code slower)

dfs_biome3 <- map(dfs_biome3, select, -matches("(occur|nfire)_(ifph|lba|mtbs)"),
                  -matches("(ifph|lba|mtbs)_prop"))
method <- names(dfs_biome3)
names(method) <- method

```

```{r}
set.seed(1234)
pred_vars <- c("afgAGB", "pfgAGB", "MAT", "MAP", "prcpPropSum")
names(pred_vars) <- pred_vars
# so that herbAGB is in the final dataframes
pred_vars2 <- unique(c(pred_vars, 'herbAGB'))

dfs_biome3_orig <- dfs_biome3 # dfs_biome is resampled below

# predictor vars are the same in both dfs
df_pred <- dfs_biome3[[1]][, pred_vars2]
```


## Bin data

20 bins strikes a good balance, where data isn't very distorted by still more
balanced than the original data. 

```{r}
n_categories <- 20 
df_pred_bin1 <- bin_df(df_pred, cols = pred_vars,
                       n_categories = n_categories)
head(df_pred_bin1)
```

## Explore bins

Number of unique bins

```{r}
n_bins <- df_pred_bin1$bin_all %>% 
  unique %>% 
  length()
n_bins
```

The 10 bins with the most 

```{r}
n_per_bin_table <- table(df_pred_bin1$bin_all)

n_per_bin <- as.vector(n_per_bin_table)

summary(n_per_bin)
names(n_per_bin) <- dimnames(n_per_bin_table)[[1]]


sort(n_per_bin, decreasing = TRUE)[1:10]
```

Histograms of the number of observations per bin

```{r }
  hist(n_per_bin, breaks = 200,
       main = 'number of observations in each multi-dimensional bin',
       xlab = "Number of observations in bin")

  hist(n_per_bin[n_per_bin < 100], breaks = 100,
       main = 'number of observations in each multi-dimensional bin \n(xlim restricted)',
       xlab = "Number of observations in bin")
```

## Resample data

```{r resample, message = FALSE}
# note this chunk of code can't be safely re-run without
# rerunning code above that creates dfs_biome3

df_pred_resample1 <-  df_pred_bin1 %>% 
  mutate(row_num = 1:nrow(.)) %>% 
  group_by(bin_all) %>% 
  slice_sample(
  # making each bin the same size so end up with
  # ~ as many rows as the original dataset
  n = round(nrow(df_pred_bin1)/n_bins),
  replace = TRUE)

# repeating the same rows in both dataframes in dfs_biome3
dfs_biome3 <- map(dfs_biome3, function(df) {
  df$bin_all <- df_pred_bin1$bin_all
  
  out <- df[df_pred_resample1$row_num, ]
  out
})

# checking that all bins have the same number of observations
dfs_biome3$paint$bin_all %>% 
  table() %>% 
  as.vector() %>% 
  unique()
```


## training data

Note--previously a small sample for now to make model fitting quicker.
Now the training dataset is all the data.

```{r train}

# rows for test dataset
# note set up this way, in case the training
# dataset is all the data
rows2sample <- 1:nrow(dfs_biome3[[1]])

# rows used in testing dataset (this is somewhat legacy code)
# and now the test dataset is mostly used to for some figures
# where smaller dataset is need (isn't actually a test dataset)
if(40000 > length(rows2sample)) {
  # if statement in case this is a test_run with few rows overall
  size = length(rows2sample)
} else {
  size <- 40000
}
rows_test <- sample(rows2sample, size = size, replace = FALSE)

# selecting rows this was so both dataframes in list have same rows
# and differences in models will only be b/ of differences in response

dfs_train <- dfs_biome3
dfs_test <- map(dfs_biome3, function(df) as.data.frame(df[rows_test, ]))

# all variables (transformed or not, that could be predictor vars)
all_pred_vars <- names(dfs_biome3[[1]]) %>% 
  str_subset(pattern = paste0("(", 
                              paste0(pred_vars2, collapse = ")|("), 
                              ")")
             )
train_pred <- dfs_train[[1]][, all_pred_vars]
test_pred <- dfs_test[[1]][, all_pred_vars]
```

# Exploratory figs & summary values


```{r summary_table}

create_summary <- function(df) {
  df %>% 
    pivot_longer(cols = everything(),
                 names_to = 'variable') %>% 
    group_by(variable) %>% 
    summarise(across(value, .fns = list(mean = mean, min = min, 
                                        median = median, max = max))) %>% 
    mutate(across(where(is.numeric), round, 4))
}

dfs_biome3[[1]][pred_vars2] %>% 
  create_summary() %>% 
  knitr::kable(caption = 'summaries of predictor variables')


response_summary <- map(dfs_biome3, function(df) {
  df %>% 
    dplyr::select(where(is.numeric), -all_of(pred_vars2), -matches("^occur_")) %>% 
    create_summary()
})

kable(response_summary$paint, 
      caption = 'summaries of response variables, calculated using paint')

```

## Plot predictor vars against each other

here using pred dataframe, because smaller and this code is slow. 

```{r pred_v_pred, cache = TRUE, dependson='train'}
ggpairs(test_pred %>% select(-matches("_")),
        lower = list(continuous = GGally::wrap("points", alpha = 0.1,    size=0.2)))
```

## boxplots-- # of fires vs predictor variables

```{r n_fires_boxplots, fig.height=9, fig.width=8, cache = FALSE, dependson='prep_data'}

# vectors of names of response variables
vars_nfire <- names(dfs_biome3[[1]]) %>% 
  str_subset("^nfire_")

vars_occur <- names(dfs_biome3[[1]]) %>% 
  str_subset("^occur_")

names(vars_occur) <- str_replace(vars_occur, "occur_", "")

vars_prop <- str_replace(vars_occur, "occur_", "") %>% 
  paste0(., "_prop")

names(vars_prop) <- names(vars_occur)

# longformat dataframes for making boxplots
dfs_biome_long <- map(dfs_biome3, function(df) {
  # for some reason select() was giving me problems
  # adding numYrs here so can take weighted average
  predvars2long(df, response_vars = c(vars_nfire, vars_prop), 
                pred_vars = pred_vars) %>% 
    mutate(across(all_of(vars_nfire), factor))
})

fire_box <- function(xvar, df, method) {
  ggplot(df, aes_string(x= xvar, y = 'value')) +
    geom_boxplot() +
    facet_wrap(~name, scales = 'free_y') +
    labs(subtitle = paste0(xvar, " (", method, " method)"))
}

map2(dfs_biome_long, names(dfs_biome_long), function(df, method) {
  map(vars_nfire, fire_box, df = df, method = method)
})

  
```

## Dotplots: deciles vs fire probability

First calculating mean of response variables (proportion of years with fires) 
for each decile of each predictor
variable. 

Next plotting deciles (originally was deciles, 10 bins, now
actually 0.5 percentiles, i.e. 200 bins)
vs fire probability. Dotplots made for each predictor
variable and response variable (as well method by which response variable
was calculated)

```{r observed_deciles}

dfs_deciles1 <- map(dfs_biome_long, longdf2deciles, response_vars = vars_prop,
                    weighted_mean = TRUE)
# df <- dfs_deciles1[[1]]
# yvar = vars_prop[[1]]

map2(dfs_deciles1, names(dfs_deciles1), function(df, method) {
  map(vars_prop, decile_dotplot, df = df, method = method,
      ylab = "fire probability (per year)")
})

```


## Histograms of response vars

Histograms for all fire variables, includes all grid cells from sagebrush biome,
showing fire probability (nfires/nyears)

```{r hists_response, message = FALSE, warning = FALSE}

dfs_prop_long <- map(dfs_biome3, function(df) {
  df %>% 
    select(matches("_prop$")) %>% 
    pivot_longer(cols = everything(),
                 names_to = "variable",
                 values_to = 'value') %>% 
    mutate(probability = value*100)
})

map(method, function(x) {
  ggplot(dfs_prop_long[[x]], aes(probability)) +
  geom_histogram(breaks = seq(-1, 15)) +
  facet_wrap(~variable, scales = "free_y") +
  labs(x = lab_fireProbPerc,
       caption = paste(x, "method used")) +
  coord_cartesian(xlim = c(-0.1, 15))
})

```

## Plot response variables against each other

```{r response_ggpairs, cache = TRUE, dependson='train'}
map(method, function(x) {
  ggpairs(dfs_test[[x]] , title =  paste0('relationships between response variables (',
                           x,  " method used)"),
        columns = vars_nfire,
        lower = list(continuous = GGally::wrap("points", alpha = 0.1, size=0.2)))
})

```

# GLMs 


## fitting models

### fitting various transforms

Creating formulas where each variable on its own is transformed numerous ways
(including formula to fit spline), all other variables are left alone,
that repeated for each variable. So have models with 1 variable transformed,
2 transformed, etc. 

see documentation for glms_iterate_transforms, in the modelling_functions.R
script

```{r iterate_transform, warning = FALSE, dependson='train', cache = FALSE}
set.seed(1234)

# adding an interaction term to help deal with over-predicting fire probability
# at pixels with high afgAGB and high MAP. parentheses around interaction
# term are included so that glms_iterate_transforms doesn't transform
# the interaction term. 
pred_vars_inter <- c(pred_vars, 'afgAGB:MAP' = "(afgAGB:MAP)",
                     #"afgAGB:prcpPropSum" = "(afgAGB:prcpPropSum)",
                     # based on examination of diagnostics plots with 
                     # Daniel, decided to try fitting this interaction
                     # (annuals and temp) as a parabola to help avoid
                     # over-predictions in warm high annual places. 
                     "afgAGB:MAT" = "(afgAGB:MAT)",
                     "afgAGB2:MAT2" = "I(afgAGB^2):I(MAT^2)",
                     "afgAGB:prcpPropSum" = "(afgAGB:prcpPropSum)",
                     "afgAGB2:prcpPropSum2" = "I(afgAGB^2):I(prcpPropSum^2)",
                     "afgAGB:pfgAGB" = "(afgAGB:pfgAGB)"#,
                     #"afgAGB2:pfgAGB2" = "I(afgAGB^2):I(pfgAGB^2)"
                     )

# pred_vars_inter <- pred_vars

mods_glm_cwf1 <- glms_iterate_transforms(
  preds = pred_vars_inter, 
  df = dfs_train$paint[ , ],
  response_var = "cwf_prop", 
  delta_aic = 10,
  weights_col = "numYrs")

# not including the last element of the list which is final_formula
mods_glm_cwf2 <- mods_glm_cwf1[-length(mods_glm_cwf1)]

map_dfc(mods_glm_cwf2, ~ names(.$aic[1:5])) %>% 
  kable(caption = "5 best transformations at each step")
best_aic_by_step <- map_dbl(mods_glm_cwf2, ~.$aic[1])

best_aic_by_step <- c(step0 = mods_glm_cwf1$step1$aic[['convert_none']], 
                      best_aic_by_step)
# AIC improvement by step
diff(best_aic_by_step)
plot(y = best_aic_by_step, 
     x = 1:length(best_aic_by_step) - 1,
     ylab = "AIC",
     xlab = "Number of variables transformed")

```

Best transformation each step.

```{r}
var_transformed <- map(mods_glm_cwf2, function(x) x$var_transformed)
var_transformed
```


### same model for all response variables

Response variables are the proportion of years in which fires occurred.
Using best model formula selected in the previous step

```{r fit_glm1, dependson='iterate_transform'}

best_form <- mods_glm_cwf1$final_formula %>% 
  str_replace(".+(?=~)", "") # remove everything before ~ (i.e. response var)
forms_glm1 <-  map(paste(vars_prop, best_form), as.formula)

print(best_form)
names(forms_glm1) <- names(vars_occur)
mod_glm1 <- map(method, function(x) {
  # fitting glms for each formula 
  map(forms_glm1, function(form) {
    glm(form, data = dfs_train[[x]], family = 'binomial', weights = numYrs)
  })
}) 

# should be the same AIC (i.e. refitting the same model)
AIC(mod_glm1$paint$cwf)

mod_glm1_flat <- flatten_rename(mod_glm1)
summary(mod_glm1_flat$paint_cwf)

```

## partial dependence & VIP

```{r pdp_glm, warning = FALSE, fig.width = 8, fig.height=8, cache = FALSE, dependson='fit_glm1'}
# vip::vip(bin_glm1) # variable importance
vip_plots1 <- map2(mod_glm1_flat, names(mod_glm1_flat), function(mod, name) {
   out <- vip::vip(mod, num_features = 15) +
    labs(subtitle = name)
})
marrangeGrob(vip_plots1, ncol = 3, nrow = 3)
map(mod_glm1_flat[1], pdp_all_vars, mod_vars = pred_vars, ylab = 'probability',
              train = train_pred)

# pdp for the original data
map(mod_glm1_flat[1], pdp_all_vars, mod_vars = pred_vars, ylab = 'probability',
              train = df_pred)
```


## observed vs. predicted

Predicting on the data
Note now that dfs_train is the whole dataset, i'm predicting on the
'training' dataframe

```{r}

# create prediction for each each model
# (i.e. for each fire proporation variable)
predict_by_response <- function(mods, df) {
  df_out <- df
  for(var in names(vars_prop)) {
    mod <- mods[[var]]
    
    response_name <- paste0(vars_prop[[var]], "_pred")
    df_out[[response_name]] <- predict(mod, df, type = 'response')
  }
  df_out
}

pred_glm1 <- map2(mod_glm1, dfs_train, predict_by_response)

dfs_test_warm <- map(dfs_train, function(df) mutate(df, MAT = MAT + 5))

pred_glm1_warm <- map2(mod_glm1, dfs_test_warm, predict_by_response)
```

```{r message = FALSE}

# long format so can have observed and predicted columns
pred_glm1_long <- map(pred_glm1, function(df) { 
  df %>% 
    select(matches("_prop")) %>% 
    mutate(row = 1:n()) %>% 
    pivot_longer(cols = -row) %>% 
    mutate(data_source = str_extract(name, "^[a-z]+"),
           type = ifelse(str_detect(name, "prop_pred$"), 'predicted', 
                         'observed')) %>% 
    select(-name) %>% 
    pivot_wider(id_cols = c('row', 'data_source'), 
                names_from = 'type',
                values_from = 'value') %>% 
    select(-row)
})

# observed vs predicted ggplots
g <- ggplot(pred_glm1_long[[1]] %>% filter(observed <= 0.1), 
            aes(y = predicted)) +
  facet_wrap(~data_source) +
  labs(subtitle = 'observed and predicted fire probabilities',
       y = "predicted probability")

g +
  geom_boxplot(aes(x = as.factor(as.numeric(as.factor(observed)) - 1)))+
  labs(x = "number of observed fires")

g +
  geom_jitter(aes(x = observed), width = 0.005, alpha = 0.1) +
  geom_smooth(aes(x = observed), method = 'lm') +
  geom_abline(slope = 1) 

```

### Deciles

Binning predictor variables into deciles (actually percentiles) and looking at the mean
predicted probability for each percentile. The use of the word deciles
is just a legacy thing (they started out being actual deciles)

Then predicting on an identical dataset but with warming

```{r}
vars_prop_pred <- paste0(vars_prop, "_pred")
response_vars <- c(vars_prop, vars_prop_pred)

pred_glm1_deciles <- map(pred_glm1, predvars2deciles, 
                         response_vars = response_vars,
                         pred_vars = pred_vars)

pred_glm1_deciles_warm <- map(pred_glm1_warm, predvars2deciles,
                              response_vars = response_vars,
                              pred_vars = pred_vars)

map2(pred_glm1_deciles, names(pred_glm1_deciles), function(df, method) {
  map(vars_prop, decile_dotplot, df = df, method = method, 
      add_predicted = TRUE)
})




# tryng
map2(pred_glm1_deciles_warm, names(pred_glm1_deciles_warm), function(df, method) {
  map(vars_prop, decile_dotplot, df = df, method = method, 
      add_predicted = TRUE, title = "data w/ +5 deg C warming")
})


```





Plot observed vs predicted values for the quantile

```{r}
# long format, so have columns of predicted and observed
# and can facet by response variable
dfs_obs_v_pred <- map(pred_glm1_deciles, function(df) {
  df %>% 
    select(matches("_prop")) %>% 
    mutate(id = 1:nrow(.)) %>% 
    pivot_longer(cols = matches("_prop")) %>% 
    mutate(type = ifelse(str_detect(name, "_pred$"),
                         "predicted", "observed"),
           variable = str_extract(name, "^[a-z]+")) %>% 
    pivot_wider(id_cols = c(id, variable), names_from = 'type')
})     

map2(dfs_obs_v_pred, names(dfs_obs_v_pred), function(df, method) {
  ggplot(df, aes(observed, predicted)) +
    geom_point() +
    facet_wrap(~variable) +
    geom_abline(slope = 1) +
    labs(title = paste(method, "method"),
         subtitle = "fire probability averaged for each decile of each predictor variable",
         caption = "each panel shows observed and predicted values
         of annual fire probability, for a given fire dataset")
})


```

R2 of observed vs predicted

```{r}
m <- lm(predicted ~ observed, dfs_obs_v_pred$paint)
summary(m)$r.squared
```

Publication quality quantile plot

```{r}

# publication quality version
g <- decile_dotplot_pq(pred_glm1_deciles$paint)

# obs/pred inset
g2 <- add_dotplot_inset(g, pred_glm1_deciles$paint)

g2
png(paste0("figures/quantile_plots/quantile_plot_v5", s,  ".png"), 
     units = "in", res = 600, width = 5.5, height = 3.5 )
g2
dev.off()
```

### Deciles--original data

Observed/predicted quantiles for the original (non resampled) dataset

```{r}

# just doing for $paint method data
pred_glm_orig <- map2(mod_glm1[1], dfs_biome3_orig[1], predict_by_response)

pred_glm_deciles_orig <- map(pred_glm_orig, predvars2deciles, 
                         response_vars = response_vars,
                         pred_vars = pred_vars)


map2(pred_glm_deciles_orig[1], names(pred_glm_deciles_orig)[1], function(df, method) {
  map(vars_prop, decile_dotplot, df = df, method = method, 
      add_predicted = TRUE, title = "Original data")
})
```



### Deciles Filtered 

20th and 80th percentiles for each climat variable

```{r}
df <- pred_glm1$paint[, c("MAT", "MAP", "prcpPropSum")] %>% 
  mutate(MAT = MAT - 273.15) # k to c
map(df, quantile, probs = c(0.2, 0.8), na.rm = TRUE)
```


Filtered 'Decile' plots of data. These plots show each vegetation variable,
but only based on data that falls into the upper and lower two deciles of
each climate variable. 


```{r glm_deciles_filtered, fig.height = 10, fig.width = 5, message = FALSE}
clim_vars <- c("MAT", "MAP", "prcpPropSum")
pred_glm1_deciles_filt <- map(pred_glm1, predvars2deciles, 
                         response_vars = response_vars,
                         pred_vars = pred_vars,
                         filter_var = TRUE,
                         filter_vars = pred_vars) 


map(pred_glm1_deciles_filt, decile_dotplot_filtered_pq)

```

Climate variables shown on the x axis

```{r fig.height = 10, fig.width = 7}
map(pred_glm1_deciles_filt, decile_dotplot_filtered_pq,
    xvars = clim_vars)
```



Publication quality


```{r insets}
df_l1 <- pred_glm1_deciles_filt$paint %>% 
  dplyr::filter(name %in% c("afgAGB", "pfgAGB")) %>% 
  split(., .$name, drop = TRUE)

# list of lists, 
df_l2 <- map(df_l1, function(df) split(df, df$filter_var))

# one list element for veg var and filter var
df_l3 <- flatten_rename(df_l2)

inset_l1 <- map(df_l3, create_inset_filt)

```



```{r, fig.gwidth = 7, fig.height = 8 }
# create inset elements

# inset locations
l <- 0.01 # left
b <- c(0.18, 0.515, 0.85) + 0.02 #bottom
w <- 0.36 # width
h <- 0.11 #height

# letters of insets correspond to the panels they belong in
A <- inset_element(inset_l1$afgAGB_MAT, l, b[3], l + w, b[3] + h)
B <- inset_element(inset_l1$pfgAGB_MAT, l, b[3], l + w, b[3] + h)
C <- inset_element(inset_l1$afgAGB_MAP, l, b[2], l + w, b[2] + h)
D <- inset_element(inset_l1$pfgAGB_MAP, l, b[2], l + w, b[2] + h)
E <- inset_element(inset_l1$afgAGB_prcpPropSum, l, b[1], l + w, b[1] + h)
f <- inset_element(inset_l1$pfgAGB_prcpPropSum, l, b[1], l + w, b[1] + h)

g2 <- pred_glm1_deciles_filt$paint %>% 
  filter(filter_var %in% clim_vars) %>% 
  decile_dotplot_filtered_pq2(insets_left = list(A, C, E),
                              insets_right = list(B, D, f),
                              ylim = c(0, 3.48))
g2
# add insets

png(paste0("figures/quantile_plots/quantile_plot_filtered_insets_v3",
            s, ".png"), units = "in", res = 600, width = 8, height = 8)
g2
dev.off()

```

version without insets 

```{r fig.width = 7, fig.height = 8}

g <- pred_glm1_deciles_filt$paint %>% 
  filter(filter_var %in% clim_vars) %>% 
  decile_dotplot_filtered_pq2()

jpeg(paste0("figures/quantile_plots/quantile_plot_filtered_v3", s,  ".jpeg"), 
     units = "in", res = 600, width = 7, height = 8 )
g 
dev.off()
g
```

Filtered quantile figure with middle 2 deciles also shown

```{r}
pred_glm1_deciles_filt_mid <- map(pred_glm1, predvars2deciles, 
                         response_vars = response_vars,
                         pred_vars = pred_vars,
                         filter_var = TRUE,
                         add_mid = TRUE)

g <- decile_dotplot_filtered_pq(df = pred_glm1_deciles_filt_mid$paint)
g

jpeg(paste0("figures/quantile_plots/quantile_plot_filtered_mid_v1", s, ".jpeg"),
     units = "in", res = 600, width = 5.5, height = 6 )
g 
dev.off()
```


### Deciles Filtered --Orig

Filtered deciles for values predicted on the original data.


```{r glm_orig_deciles_filtered, fig.height = 6}

# just runningfor the paint method data 
pred_glm_orig_deciles_filt <- map(pred_glm_orig[1], predvars2deciles, 
                         response_vars = response_vars,
                         pred_vars = pred_vars,
                         filter_var = TRUE)

map2(pred_glm_orig_deciles_filt, names(pred_glm_orig_deciles_filt), function(df, method) {
  map(vars_prop, decile_dotplot_filtered, df = df, method = method,
      title = "Original (not resampled) data")
})

```

# Save output

```{r save_output}
# glm models
mods2save <- mod_glm1_flat
mods2save$formula <- best_form
mods2save$pred_vars_inter <- pred_vars_inter # so have interactions
saveRDS(mods2save, 
        paste0("models/glm_binomial_models_byNFire_v2_bin",
               n_categories, "_cwf", s, ".RDS"))

```

# session info

Hash of current commit (i.e. to ID the version of the code used)

```{r}
system("git rev-parse HEAD", intern=TRUE)
```

Packages etc.

```{r}
sessionInfo()

```
