==> scripts/05_resampled-data_examine.Rmd <==
---
title: "Resampling observed fire probability data from the sagebrush biome"
author: "Martin Holdrege"
date: "`r lubridate::today()`"
output:
  html_document:
    toc: true
    toc_float:
      collapsed: false
---

Here the training data is resampled to be more balanced (i.e., upsampling
bins of rare data, downsampling bins of common data). 

Note a concern with the up-sampling is that many bins start with only one
datapoint in them, which is then repeated many times in the resampled datset.
Apparently, having duplicated points like this can increase the risk of 
over-fitting. 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      message = FALSE)
```

# Dependencies 

Sourcing the script that creates a dataframe of historical
fore occurrence, as well as biotic and climate
predictors for those same grid-cells. Here the biomass data
is different before and after a fire (i.e. can have multiple
rows of data per cell)

```{r source_data, warning=FALSE, message=FALSE}
source("scripts/04_create_biome-mask_dataframe_byNFire.R")
```

```{r warning=FALSE, message=FALSE}
source("src/general_functions.R")
source("src/modeling_functions.R")
source("src/resample_functions.R")
source("src/basemaps.R")
theme_set(theme_classic())
```


# Prep data

```{r prep_data}
dfs_biome3 <- dfs_byNFire3
method <- names(dfs_biome3)
names(method) <- method

#dfs_biome3 <- map(dfs_biome3, slice_sample, n = 1000) # for testing

```


```{r}
set.seed(1234)
pred_vars <- c("afgAGB", "pfgAGB", "MAT", "MAP", "prcpPropSum")
names(pred_vars) <- pred_vars
# so that herbAGB is in the final dataframes
pred_vars2 <- c("afgAGB", "pfgAGB", 'herbAGB', "MAT", "MAP", "prcpPropSum")
names(pred_vars2) <- pred_vars2
# predictor vars are the same in both dfs
df_pred <- dfs_biome3[[1]][, pred_vars2]
```


## Bin data

```{r}
dfs_bin <- list()
bin_values <- c(5, 10, 20, 29)
bin_names <- paste0("bin", bin_values)

names(bin_values) <- bin_names
dfs_bin <- map(bin_values, function(x) {
  bin_df(dfs_biome3$paint, cols = pred_vars,
                        n_categories = x)
})

```

## Explore bins

Number of unique bins

```{r}
n_bins <- map_dbl(dfs_bin, function(df){ 
  df$bin_all %>% 
    unique %>% 
    length()
  })
n_bins

```

Percent of total possible number of bins filled

```{r}
(n_bins/(bin_values^5)*100) %>% 
  round(2)
```


Number of observations per bin

```{r}
n_per_bin <- map(dfs_bin, function(df) {
  x <- table(df$bin_all)
  
    out <- as.vector(x)
  names(out) <- names(x) # naming each vector element by the name of the bin
  out
})

map(n_per_bin, summary)
```

Percent of bins with 1 observation

```{r}
map_dbl(n_per_bin, function(x) mean(x ==1)*100)

```

# Map of bin density

```{r}

df_cell_bin <- dfs_bin$bin20 %>% 
  select(cell_num, bin_all) %>% 
  # adding in the number of observation for a given bin_all
  mutate(bin_size = n_per_bin$bin20[bin_all]) %>%
  # multiple observations can occur per cell (when have >0 fires)
  # so taking the average bin size for observations belonging
  # to the given gridcell
  group_by(cell_num) %>% 
  summarise(bin_size = mean(bin_size))

```

fill raster with bin sizes

```{r}

r_bin <- rast_cwf1
r_bin[] <- NA
r_bin[df_cell_bin$cell_num] <- df_cell_bin$bin_size

cutoff <- resample_bins(dfs_bin$bin20, return_target_bin_size = TRUE)

# percent of grid cells belonging to bins 
mean(df_cell_bin$bin_size < cutoff)*100
r_cutoff <- r_bin
r_cutoff[] <- NA_integer_
r_cutoff[r_bin < cutoff] <- 1L
r_cutoff[r_bin == cutoff] <- 2L
r_cutoff[r_bin > cutoff] <- 3L


```

Create map of bin sizes

```{r}
breaks <- c(1, 4, 9, 20, 50, 100, 500, 1000, 10000)
tm_bin <- tm_shape(r_bin) +
  tm_raster(
    breaks = breaks,
    labels = label_creator(breaks),
    palette = "GnBu",
    title= "Observations/bin"
  ) +
  base

tm_bin

jpeg("figures/maps_resampling/bin20_density_v1.jpeg", res = 600, 
     height = 4, width = 4, units = "in")
tm_bin 
dev.off()


```

Map of whether a location was upsampled or downsampled

```{r}
tm_cutoff <- tm_shape(r_cutoff) +
  tm_raster(
    breaks = seq(0.5, 3.5, by = 1),
    labels = c("upsampled", "no change", "downsampled"),
    palette = c("#67a9cf", "gray", "#ef8a62"), # from RdBu palette
    title =""
  ) +
  base
tm_cutoff
```


# Resample data

Re-sampling multiple times, to get a sense for the variability
between random draws

```{r resample, message = FALSE}

samples <- letters[1:3]

dfs_resampled <- list()

for(x in samples) {
  dfs_resampled[[x]] <- map(dfs_bin, resample_bins)
}

dfs_resampled <- flatten_rename(dfs_resampled)

```


# Quantile plots

Calculating mean values per percentile of each predictor variable, for both
the original data and the resampled data

```{r}

response_vars <- "cwf_prop"
deciles_obs <- predvars2deciles(dfs_biome3$paint, 
                               response_vars = response_vars,
                               pred_vars = pred_vars) %>% 
  mutate(resample = "Original data")

deciles_resampled1 <- map_dfr(dfs_resampled, predvars2deciles,
                         response_vars = response_vars,
                         pred_vars = pred_vars,
                         .id = 'resample')

deciles_comb <- bind_rows(deciles_obs,
                          deciles_resampled1) %>% 
  mutate(replicate = str_extract(resample, "^[a-z](?=_)"),
         type = str_replace(resample, "^[a-z]_", ""))
deciles_comb %>% 
  arrange(mean_value)

```

creating the plots.

Note multiple lines per type, shows the multiple random samplings,
which are all almost identical. 

```{r, fig.height=8, fig.width = 9}

# colors--I don't like the lightest color, so not using it
cols <- RColorBrewer::brewer.pal(length(dfs_bin) + 1, "YlOrBr")[-1]
names(cols) <- names(dfs_bin)
cols <- c("Original data" = "black",
          cols)

decile_base <- function(x) {
  list(    
    facet_wrap(~name, scales = 'free_x'),
  labs(x = "mean of quantile of predictor variable",
       y = "Observed fire probability (cwf)",
       subtitle = "Original and resampled data",
       caption = "bin(n) refers to the number of bins each
       predictor variable was split into"),
  theme(legend.position = 'top',
        legend.title = element_blank()),
  scale_color_manual(
    values = cols))

}

g <- ggplot(deciles_comb, aes(x = mean_value, y = cwf_prop,
            color = type, group = resample)) +
   geom_smooth(se = FALSE, size =0.75) +
   decile_base() 
  
g
```

Decile plots showing points for just a single bin size at a time

```{r, fig.height=7, fig.width = 8}

bin_names_a <- paste0("a_", bin_names)
for(x in bin_names_a) {
  print(x)
  df <- deciles_comb %>% 
    filter(resample %in% c("Original data", x))
  
  g <- ggplot(df, aes(x = mean_value, y = cwf_prop,
            color = type, group = resample)) +
   geom_smooth(se = FALSE, size =0.75) +
    geom_point() +
    decile_base() 
  
  print(g)
  
}

```




# Histograms

Combining into a single dataframe. (only using one random sample of
the bin resampling)

```{r }

# grabbing just the first random sample
df_comb <- dfs_resampled[bin_names_a] %>% 
  # selecting certain rows just resulting dataframe is smaller
  map(function(df) df[pred_vars2]) %>% 
  bind_rows(.id = "resample") %>% 
  mutate(resample = str_replace(resample, "a_", "")) %>% 
  bind_rows(dfs_biome3$paint[pred_vars2] %>% 
              mutate(resample = "Original data")) %>% 
  mutate(resample = factor(resample, levels = names(cols)))

comb_means <- df_comb %>% 
  group_by(resample) %>% 
  summarise(across(all_of(pred_vars2), mean))
  
```

Histograms of each predictor variable, 

```{r histograms, fig.height = 8}
map(pred_vars2, function(x) {
  g <- ggplot(df_comb) +
    geom_histogram(aes_string(x = x), bins = 100) +
    facet_wrap(~resample, ncol = 2, scales = 'free_y') +
    geom_vline(data = comb_means, aes_string(xintercept = x),
               linetype = 2) +
    labs(title = x,
        subtitle = "Histograms of original and resampled data",
         caption = "Vertical line shows the mean value")
  g
})
```

### bin frequency

Plots showing how many times each of the bins of a given variable
occurred in the re-sampled datasets.

```{r fig.width = 8, fig.height = 6}

bins_comb <- dfs_resampled[bin_names_a] %>% 
  # selecting certain rows just resulting dataframe is smaller
  map(function(df) df[, "bin_all"]) %>% 
  bind_rows(.id = "resample") %>% 
  mutate(resample = str_replace(resample, "a_", "")) %>% 
  separate(bin_all, into = pred_vars, sep = "_") %>% 
  mutate(across(-resample,
                function(x) str_extract(x, "\\d+$"))) %>% 
  mutate(across(-resample, 
                function(x) factor(as.numeric(x))),
         resample = factor(resample, levels = bin_names)
         )

map(pred_vars, function(x) {
  ggplot(bins_comb, aes_string(x)) +
    geom_bar() +
    facet_wrap(~resample, drop = TRUE, scales = "free") +
    labs(x = paste("bin of", x),
         title = x,
         caption = "Number of occurrences of each bin in the resampled
         dataset")
})

```

# Pairwise comparisons

```{r}
pairwise_vars <- expand_grid(x = pred_vars,
                             y = pred_vars) 

# keeping only the rows so that have unique combos of x and y
# (e.g., so not plotting afgAGB ~ pfgAGB & pfgAGB ~ afgAGB)
rows2keep <- matrix(1:25, ncol = 5) %>% 
  lower.tri(diag = FALSE) %>% 
  as.vector()

pairwise_vars <- pairwise_vars[rows2keep, ]

pairwise_vars %>% 
  knitr::kable()

```

## GAMs

```{r gams, eval=TRUE, cache=TRUE}
# something about delay evaluation (maybe), is causing
# aes_string not to work when filling a list 
for(i in 1:nrow(pairwise_vars)) {
  
  x <- pairwise_vars$x[i]
  y <- pairwise_vars$y[i]

  g <- ggplot(df_comb, aes_string(x, y)) +
    geom_smooth(aes(color = resample), se = FALSE) +
    scale_color_manual(values = cols)
  
  print(g)

}
```


## 2d density

```{r density_plots, fig.height=7, fig.width = 9}
for(i in 1:nrow(pairwise_vars)) {
  
  x <- pairwise_vars$x[i]
  y <- pairwise_vars$y[i]

  g <- ggplot(df_comb, aes_string(x, y)) +
    geom_density_2d(bins = 25) +
    facet_wrap(~resample, scales = 'free') 
    
  print(g)

}
```
